+++ test/resources/access-assignment-doc-definitions.js
@@ -6,17 +6,25 @@
     },
     accessAssignments: [
       {
+        type: 'channel',
         users: 'user3',
         channels: 'channel3'
       },
       {
+        type: 'channel',
         roles: 'role3',
         channels: 'channel4'
       },
       {
+        // With no "type" property defined, it should default to channel access assignment
         users: [ 'user1', 'user2', 'user4' ],
         roles: [ 'role1', 'role2' ],
         channels: [ 'channel1', 'channel2' ]
+      },
+      {
+        type: 'role',
+        users: 'user5',
+        roles: 'role4'
       }
     ]
   },
@@ -35,8 +43,18 @@
     },
     accessAssignments: [
       {
+        type: 'role',
         users: function(doc, oldDoc) {
           // The sync function template should automatically replace the oldDoc param value with null if its _deleted property is true
+          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.users;
+        },
+        roles: function(doc, oldDoc) {
+          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.roles;
+        }
+      },
+      {
+        type: 'channel',
+        users: function(doc, oldDoc) {
           return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.users;
         },
         roles: function(doc, oldDoc) {
@@ -47,6 +65,7 @@
         }
       },
       {
+        // With no "type" property defined, it should default to channel access assignment
         roles: function(doc, oldDoc) {
           return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.roles;
         },
@@ -55,6 +74,7 @@
         }
       },
       {
+        // With no "type" property defined, it should default to channel access assignment
         users: function(doc, oldDoc) {
           return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.users;
         },